fix conversion to logical cell array issue

+++ a5-moo/Mutate.m
@@ -4,7 +4,7 @@ function children = Mutate(~, children, mutationRate)
 numGene = length(children(1).Genome);
 numGenome = length(children);
 genomes = reshape(extractfield(children, 'Genome'), numGene, numGenome)';
-childrenGenomes = GeneticEncoding.BinaryOperators.Mutate(genomes, mutationRate);
+childrenGenomes = double(GeneticEncoding.BinaryOperators.Mutate(genomes, mutationRate));
 
 parfor i = 1:numGenome
     children(i).Genome = childrenGenomes(i, :);
